기능 추가

+++ widgetslib/src/main/java/nah/prayer/widgetslib/corners/CornerButton.kt
@@ -14,7 +14,7 @@ class CornerButton(context: Context, attrs: AttributeSet?) : AppCompatButton(con
         this.background = styleShorten.setStyle(context, attrs, CornersEnum.C_BUTTON)
     }
 
-    fun setNaStrokeSize(width:Float){
+    fun setNaStrokeSize(width:Int){
         this.background = styleShorten.setCustomStrokeSize(width)
     }
